web_app/back_end/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
awscli==1.32.79
boto3==1.34.79
certifi==2024.2.2
charset-normalizer==3.3.2
django-cors-headers==4.3.1
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
joblib==1.3.2
keras==2.15.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes~=0.2.0
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
protobuf==4.23
Pygments==2.17.2
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.13.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.15
tensorboard-data-server==0.7.2
tensorflow==2.15.0
termcolor==2.4.0
threadpoolctl==3.4.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.14.0
yfinance==0.2.37
